requirements.txt
# requirements.txt
#flask==2.0.3 novo
flask[async]
Werkzeug==2.0.3
requests==2.26.0
python-dotenv==0.19.0
aiohttp --only-binary=:all: # Para requisições assíncronas novo
asyncio==3.4.3 # Para processar múltiplas requisições novo
redis-5.2.1 #novo